+++ rest/src/decorators/index.ts
@@ -2,6 +2,8 @@ export * from './errorPassThrough.decorator';
 export * from './errorPassThroughHttp.decorator';
 export * from './handle400WithValidations.decorator';
 export * from './handle404.decorator';
+export * from './httpClientErrorsMapper.decorator';
+export * from './httpClientValidationErrorsMapper.decorator';
 export * from './ignoreClientErrors.decorator';
 export * from './suppressError.decorator';
 export * from './withRestClientError.decorator';