+++ rest/src/middlewares/index.ts
@@ -1,1 +1,2 @@
 export * from './clientErrorHandling.middleware';
+export * from './httpClientErrorProcessing.middleware';