+++ rest/src/misc/defaults.ts
@@ -15,7 +15,7 @@ export const ERROR_HANDLING_REST_MIDDLEWARES_ORDER: RestMiddlewareOrderType<midd
     'QueryParameterMiddleware',
     'HeadersMiddleware',
     'HeaderParameterMiddleware',
-    'ClientErrorHandlingMiddleware',
+    'HttpClientErrorProcessingMiddleware',
     'ProducesMiddleware',
     'LoggerMiddleware',
     'IgnoredInterceptorsMiddleware',