+++ src/errorHandlers/handle404Func.ts
@@ -10,14 +10,14 @@ import {Handle4xxOptions, XXXHttpClientError} from '../interfaces';
  * @param errorReturnCallback - Callback that transforms unprocessed http error response into TError
  * @param clientErrorReturnCallback - Callback that transforms processed http error response error messages into TClientError
  */
-export function handle404Func<TError, TClientError>(error: HttpErrorResponse,
-                                                    options: Handle4xxOptions,
-                                                    errorReturnCallback: (error: HttpErrorResponse) => TError,
-                                                    clientErrorReturnCallback: (error: XXXHttpClientError) => TClientError): TError|TClientError
+export async function handle404Func<TError, TClientError>(error: HttpErrorResponse,
+                                                          options: Handle4xxOptions,
+                                                          errorReturnCallback: (error: HttpErrorResponse) => TError,
+                                                          clientErrorReturnCallback: (error: XXXHttpClientError) => TClientError): Promise<TError|TClientError>
 {
     if(error.status == 404)
     {
-        const {errors} = readErrorsFromHttpErrorResponse(error, options.injector, options.clientErrorsResponseMapper);
+        const {errors} = await readErrorsFromHttpErrorResponse(error, options.injector, options.clientErrorsResponseMapper);
 
         return clientErrorReturnCallback(
         {