+++ src/errorHandlers/handleHttpClientErrors.ts
@@ -7,12 +7,12 @@ import {HttpClientErrorOptions} from '../interfaces';
 /**
  * Handles http error response with code 400..499 and returns HttpClientError, otherwise returns original http error response
  */
-export function handleHttpClientErrors(error: HttpErrorResponse, options: HttpClientErrorOptions): HttpClientError|HttpErrorResponse
+export async function handleHttpClientErrors(error: HttpErrorResponse, options: HttpClientErrorOptions): Promise<HttpClientError|HttpErrorResponse>
 {
     //handles client codes (400..499)
     if(error.status >= 400 && error.status < 500)
     {
-        const {errors, validationErrors} = readErrorsFromHttpErrorResponse(error, options.injector, options.clientErrorsResponseMapper, options.clientValidationErrorsResponseMapper);
+        const {errors, validationErrors} = await readErrorsFromHttpErrorResponse(error, options.injector, options.clientErrorsResponseMapper, options.clientValidationErrorsResponseMapper);
 
         return new HttpClientError(errors, validationErrors, error);
     }