+++ src/misc/types.ts
@@ -1,17 +1,17 @@
 import {HttpErrorResponse} from '@angular/common/http';
-import {Func1, Func4Rest} from '@jscrpt/common';
+import {Func1, Func4Rest, PromiseOr} from '@jscrpt/common';
 
 import {Handle4xxOptions, XXXHttpClientError, HttpClientValidationErrors} from '../interfaces';
 
 /**
  * Response mapper for http client errors that are converted to array of error messages
  */
-export type HttpClientErrorResponseMapper = (err: HttpErrorResponse) => string[];
+export type HttpClientErrorResponseMapper = (err: HttpErrorResponse) => PromiseOr<string[]>;
 
 /**
  * Response mapper for http client validation errors that are converted into object RestClientValidationErrors
  */
-export type HttpClientValidationErrorResponseMapper = (err: HttpErrorResponse) => HttpClientValidationErrors|null;
+export type HttpClientValidationErrorResponseMapper = (err: HttpErrorResponse) => PromiseOr<HttpClientValidationErrors|null>;
 
 /**
  * Custom handler for `HttpErrorResponse`